+++ ZakLyd.Web.InterfaceEntities/InterfaceHotelSummary.cs
@@ -2,7 +2,7 @@
 {
     public class InterfaceHotelSummary
     {
-        public int HotelId { get; set; }
+        public long HotelId { get; set; }
         public string HotelName { get; set; }
         public int Stars { get; set; }
         public int DistanceFromHaram { get; set; }
